Fix some items acting as uninteractable watermelons
The issue is that upon item entity spawn, the entity is always set to watermelon as a default. Then, in ENT:setItem, that code is run again, this time with the proper item model. Most of the time that's fine, however some models have funky physics that dont update, causing the model to stay with the watermelon physics, and not not traceable. Here we just set the correct model right away, avoiding that issue