Add a radius property to allow a dynamic chat range
As things stand, it's not possible for a range to be bound to a dynamic variable (as for example with the "chatRange" parameter which is only retrieved during registration). In order not to break compatibility with commands created on other addons/servers, I introduce a new property as a function and under the name "radius". This function will return the value of a parameter each time it's called, thus correcting this little fixed range issue.