Fixed Vendor Crash Exploit
If the player doesn't have space on their inventory, the item would be dropped. 

This would increasingly hurt the server performance by overflowing the net channels, which can eventually lead to a crash.

This exploit has been here since 1.1 and it's time it gets fixed.

This fix consists in refunding the money and warning the player that there is no space for such item.